pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "daily-etf-analysis"
version = "0.2.0"
description = "A modern Python project template with loguru, ruff, and uv"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Tendo33"},
]
maintainers = [
    {name = "Tendo33"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
keywords = ["template", "python", "loguru", "ruff", "uv"]
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.0.0",
    "aiofiles>=23.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "akshare>=1.12.0",
    "efinance>=0.5.5",
    "yfinance>=0.2.0",
    "exchange-calendars>=4.5.0",
    "litellm>=1.80.10",
    "tiktoken>=0.8.0,<0.12.0",
    "json-repair>=0.55.1",
    "PyYAML>=6.0",
    "tavily-python>=0.3.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
    "mypy>=1.15.0",
    "vulture>=2.10",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.23.0",
    "anyio>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/Tendo33/daily-etf-analysis"
Repository = "https://github.com/Tendo33/daily-etf-analysis"
Documentation = "https://github.com/Tendo33/daily-etf-analysis/tree/main/doc"
Changelog = "https://github.com/Tendo33/daily-etf-analysis/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/Tendo33/daily-etf-analysis/issues"

# Ruff configuration
[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".agent",
    ".claude",
    ".codex",
    ".cursor",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.10+
target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["E501"]  # Line length is handled by formatter

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
known-first-party = ["daily_etf_analysis"]

[tool.mypy]
python_version = "3.10"
files = ["src"]
mypy_path = "src"
warn_unused_configs = true
show_error_codes = true
pretty = true
disable_error_code = ["import-untyped"]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "asyncio: marks tests as async tests",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# Coverage configuration
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
skip_covered = false

# Hatch configuration for building
[tool.hatch.version]
path = "src/daily_etf_analysis/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = [
    "/.agent",
    "/.claude",
    "/.codex",
    "/.cursor",
    "/.cursorrules",
    "/.github",
    "/data",
    "/doc",
    "/dist",
    "/tests",
    "/scripts",
    "/examples",
    "/htmlcov",
    "/logs",
    "/skills",
    "/uv.lock",
    "*.md",
]

[tool.hatch.build.targets.wheel]
packages = ["src/daily_etf_analysis"]
